Update patch set 1
I would prefer that you didn't submit this.

The power management code already sends an Intent.ACTION_BATTERY_CHANGED
broadcast intent when the plug state changes, and that broadcast
includes a description of the new plug state.  It should not be sending
redundant broadcasts (and so does not need these additional Intent
fields).

Patch-set: 1