Update patch set 1
No score.

The goal of this is to enable a number of scenarios for applications
that are not currently awake to be able to receive AC power
notifications.

The ACTION_BATTERY_CHANGED broadcast is set to only broadcast to already
registered receivers (i.e. only ones already opened).  This makes sense
since the broadcast can trigger rather frequently since it reports
battery charge state.  It would be bad for applications to constantly
get woken up for that.

The new ACTION_POWER_CONNECTED/DISCONNECTED broadcasts are far more
occasional, and so it's okay for an application to be woken.

Patch-set: 1