Update patch set 1
Looks good to me, but someone else must approve.

This works for me. I realise that this simply moving code, but there
appears to be confusion between TARGET_DEVICE, and TARGET_PRODUCT.

As I understand it, you can have multiple products using the same
device, so makes sense for this to be using TARGET_DEVICE, rather than
TARGET_PRODUCT, but probably this isn't the patch to do that in.

Patch-set: 1