Update patch set 1
Do not submit.

We should split up the build system changes from the add a new target
changes.  Also, the build system change you're reverting was the result
of a bunch of discussion before it was put in and I think it does some
useful stuff -- might be worth not just undoing it.

For the add a new target changes, they should end up in a
platform/vendor/openmoko/freerunner tree (or whatever the appropriate
vendor name is), rather than in the main build.
See http://android.git.kernel.org/?p=platform/vendor/htc/dream.git for
an example.

Was it necessary to change the prelink mappings?  We may have to make
those per-target.

Patch-set: 1