Update patch set 1
No score.

My apologies.
Previous post was incorrect.

There was update to build system with drops from cupcake, now
envsetup.sh runs core/config.mk, which is not setting shell to bash (it
is done in core/main.mk), and affected by this change script is executed
with default shell - sh.

So your change is correct with '-a'.
However '[[' must be  changed to '[' anyway to be sh compliant.

PS:
I'm not sure it's expected behavior to interpret script with sh, cause
core/main.mk redefines SHELL to /bin/bash and includes core/config.mk.
So, bash will be used while building process, but typing 'lunch' (or
other command, that results in requests for build variables via
config.mk) in console will cause  usage of sh for the same script. It's
weird.

Patch-set: 1