Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

When I originally put this forward my target was non-thumb (due to compiler issues bugs in armv4t), and I did at that time also have patches for armv4 bionic and dalvik, however these never got to a clean enough point to be merged.

I think at this stage supporting a non-thumb target would be at best an academic exercise, so I support this change. 

My only change to this would be placing a comment in the else clause noting this decision to not support thumb targets so that we ensure no-one else goes down this unproductive path again.

Patch-set: 1