Update patch set 2
It seems ANDROID_MAKER is undefined when you start this file, definition of it was changed later (chooseMaker() is not used anymore). This change is dependency of other changes for envsetup.sh. I'll try to collect all changes for envsetup.sh in one change and abandon this and other partial changes for envsetup.sh

Patch-set: 2