Update patch set 2
Patch Set 2: Abandoned

After much discussion, we've determined that we're not going to be able to support building Android on FreeBSD. We're worried that accepting those changes could make a device manufacturer mistakenly believe that FreeBSD is a supported host environment, with unpredictable results. To avoid such risks, we're unable to accept FreeBSD-related changes.

We thank you very much for your enthusiasm, for your patience, and for your interest in Android.

Patch-set: 2