Update patch set 5
Patch Set 5:

it's because the way the framework handles the kl/kcm files. They have to match the input device name, and on x86, the generic AT keyboard shows up as that. I think the qwerty keyboard was missing some keys (like Menu, Alt, or something like that).

Maybe none of this is no longer necessary?

Patch-set: 5