Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Note: we don't normally make changes to the open-source eclair branch, except in the cts project. Once this change is approved, I'll need to get involved to properly merge it.

In the longer term, it'd be good to migrate more of the cts-related files under the cts project.

Patch-set: 1
Label: Code-Review=+1