Update patch set 1
Patch Set 1:

That's hardly extensible.
I'd recommend taking the default ones OUT of envsetup.sh.

BTW, we also need a way to deal with the TARGET_ARCH set properly from a lunch order. Right now, we have to set this in $TOP/buildspec.mk... not nice.

If you really want to move the call to envsetup.sh... OK... but it doesn't seem right to me.

Patch-set: 1
Label: Code-Review=0