Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

There hadn't been any use case for this before, and the only use case we found for this flag (fast-math everywhere except in sqlite) turned out to not be appropriate to justify a near-system-wide flag change.

Because the build system is already so complex, I'd really prefer to avoid introducing unused features.

Unless we hear otherwise, we'll be abandoning this change in the next few weeks.

Patch-set: 1