Update patch set 2
Patch Set 2:

Yes, I already caught the assumption via the prebuilt.mk. The original prebuilt.mk requires PRESIGNED. 

So I modified the prebuilt.mk just for the APKs which is not marked with PRESIGNED following the assumption, because I needed to sign an prebuilt APK (which is prebuilt, but not signed) with a platform key.

Is this case not acceptable really?

Patch-set: 2