Update patch set 2
Patch Set 2:

You have to sign it before putting it into LOCAL_SRC_FILES.
How do you build an apk without signing it? If you build it with the android platform build system, it should be signed with some of the platform keys. Otherwise you can sign it manually.

Patch-set: 2