Update patch set 1
Patch Set 1: Verified

Style OK, merges and builds fine.

As far as I know it does what it says it does (which isn't hard since it's a one-liner). I'm guessing that we won't be able to guarantee that it'll continue working forever as it might evolve or disappear if/when we refactor the build system.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=+1