Update patch set 1
Patch Set 1:

Yeah, I agree with Ying.  The definitions.mk file is the very core of the build system.  I don't like the idea of adding a facility for it to be modified based on which vendor trees are installed.

There is probably a better way of factoring out whatever you're trying to do.

Patch-set: 1