Update patch set 2
Patch Set 2:

It is possible to get prebuilt but unsigned APKs. I know that I can sign them manually, but this patch aims for reducing the manual signing. It was quite a burden for me, as the number of packages were increased. Thanks a lot for reviewing this stuff.

Patch-set: 2