Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I've been wondering about the whole "first makefile under" thing... I'm using it now for a directory in prebult. If you have the structure:

  a
b c d

where c and d each have an Android... but a does not, the fixed routine will find the a/c/Android.mk (deterministically). a/d/Android.mk will never be found/used in this case.

Is this, therefore, an inappropriate use of the routine and should there always be a a/Android.mk if there are makefiles below it?

Patch-set: 1
Label: Code-Review=+1