Update patch set 1
Patch Set 1:

yes, we can modify the generating script as well. Most importantly meta-data needs to also have this which will allow the server if needed to validate. 
This is in response to the recent incident of the GOTA setup which took us a while to determine the cause due to the pre-timestamp being greater than post-timestamp

Patch-set: 1