Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I'm putting this patch temporarily on hold because it modifies the location of SDK platform images in the SDK installation path, and I don't know yet if we're ok with doing this in the gingerbread branch. I've added xav and raphael to discuss this.

Normally, the tools will come from the AOSP platform/sdk.git repository which already contain the code to deal with it, but I'd like their confirmation before accepting it.

Patch-set: 1
Reviewer: David Turner <1000411@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0