Update patch set 3
Patch Set 3:

Since I find this change so weird, I'd like you to do me a favor.

Change your code so that it prints a big fat warning whenever your indexOf method disagrees with the built-in indexOf method. You'll need to call both methods, but that's okay for a test. The warning should include all of the input strings.

Then run the code as many times as necessary for this warning to show up. If you can do this and demonstrate that there's a bug in Oracle Java's indexOf() method, I'll +2 this change and we'll also report a very high priority bug to them.

Patch-set: 3