Update patch set 1
Patch Set 1:

I don't think it will cause any fundamental harm to have a language on the list that doesn't have translations, but I don't think this will actually show up in the locale picker without a country as part of the locale code.

I wish I could remember exactly why that is the case, but early on there was something on the backend that was unhappy when it got just "de" as the locale for Germany (the first translation) instead of "de_DE", so the locale picker enforces that there must be a country.

Patch-set: 1
Label: Code-Review=0