Update patch set 3
Patch Set 3:

It might work very well for your use case, but it's possible that some more complex ones wouldn't work. I'm pretty sure that the builds for some of the flagship devices are more than twice as large as the AOSP ones, and I'm pretty sure that some OEM builds are 4 or 5 times as large as AOSP builds. The risk of regression isn't worth trying to go against Moore's law.

Patch-set: 3