Update patch set 1
Patch Set 1:

I'm having trouble finding how to do the change with repo so if you can do the change that'd be much appreciated. 

Making the change with git commit --amend and then repo upload ends up with repo saying there are no branches to upload, the docs at http://source.android.com/source/using-repo.html say use --replace on repo upload, but --replace seems to not be a valid option any more, and so I took it to mean that because it wasn't supported in git it wasn't possible via repo.

Patch-set: 1
Label: Code-Review=0