Update patch set 1
Patch Set 1:

Hi JBQ:


The elfutils is change to GPLv2 with exception (allow to link with Apache 2 license) since v0.120 and the current version in AOSP is v0.97.


So it's seem this program should be GPLv2 since the elfutil in AOSP source tree is still under GPLv2...


Any suggestion? update elfutils or re-write part of read symbol?

Patch-set: 1