Update patch set 1
Patch Set 1:

For better understanding of the licensing issues Kito mentioned, you can refer to the past discussion:
    http://lists.linaro.org/pipermail/linaro-dev/2011-June/005490.html

The latest elfutils already clarifies the licensing with the following explicit exception terms:
    https://fedorahosted.org/elfutils/browser/EXCEPTION

"This file describes the limits of the Exception under which you are allowed to distribute Non-GPL Code in linked combination with Red Hat elfutils.  For the full text of the license, please see one of the header files included with the source distribution or the file COPYING found in the top level directory of the source."

However, the root cause of Aprof's GPL enforcement resulting from the fact that AOSP still uses the older version of libelf, which has no explicit exception notice.

Therefore, elftree (source: development/elftree/, Apache License in original claim) has a potential licensing problem as Aprof since it links to libelf statically.

Patch-set: 1
Label: Code-Review=0