Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The original patch from Ying Wang to enforce the 10.6 SDK was intentionally doing so on 10.7. This is due to incompatibilities between the 10.7 and 10.6 SDKs shown up by developers writing code which worked on the 10.7 SDK but failed on the 10.6 based build machines.

This patch, although useful for 10.7 based folk, would go against that intention and re-introduce the same problem Ying was solving and therefore I can't recommend its' inclusion.

I am working with Apple to see what other options are available for the 10.6 SDK in Xcode 4.3 to solve the issues.

Patch-set: 1
Label: Code-Review=-1