Update patch set 2
Patch Set 2:

Patch set 2: change compiler check to hash the compiler binary. I ran a few tests and the build time difference (if any) was lost in the variance between runs.

About GPLv3: what part of GPLv3 would be problematic? Since ccache is only used in the build and doesn't go on the device, I can't imagine it would run afoul of the anti-Tivoization clauses in GPLv3.

What I did locally was to add a check in the build to see if the host workstation already had ccache installed, and if so use that instead of the prebuilt 2.4 in the repo. If the fact that ccache 3.x is GPLv3 becomes a dealbreaker I could clean up and submit this logic.

Patch-set: 2