Update patch set 1
Patch Set 1: Verified

thanks bdc!

kroot: the reason this worked for me is that if you build manually in bionic (as i had to while coming up with the new file), then "adb sync", this file ends up on your device. it's just not in the system image. that's what this change is for.

i'll be back later (today?) to remove the three old files...

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: SUBM=+1
Label: Verified=+1