Update patch set 1
Patch Set 1:

Has this been checked on the Google build servers? 

The main reason for the BUILD_MAC_SDK_EXPEREMENTAL flag was developers creating code with the 10.7 / 10.8 SDK which failed to build on the internal build servers which were 10.6 based. The patch looks good to me, but if it's going to cause issues with the internal build servers then it's not usable as-is and we'll need the BUILD_MAC_SDK_EXPEREMENTAL flag back.

Patch-set: 1
Label: Code-Review=0