Update patch set 1
Patch Set 1:

you're misunderstanding the zlib FAQ entry. Deflate64 is unrelated to Zip64. the Deflate64 "64" is the 64KiB dictionary it uses instead of the regular 32KiB one. the Zip64 "64" is the number of bits used to represent various sizes. you can (and usually do) have Zip64 files with regular Deflate entries.

Patch-set: 1