Update patch set 5
Patch Set 5:

[https://blogs.oracle.com/darcy/entry/bootclasspath_older_source]

"[...]
To use javac from JDK N to cross-compiler to an older platform version, the correct practice is to:
- Use the older -source setting.
- Set the bootclasspath to compile against the rt.jar (or equivalent) for the older platform.
If the second step is not taken, javac will dutifully use the old language rules combined with new libraries,
which can result in class files that do not work on the older platform since references to non-existent methods can get included.
[...]"

In JDK 7, a new warning is raised to warn about the use of an older platform version without the "bootclasspath" argument.

When we compile with the JDK 6, this situation is also detected as suspicious but the user is not warned about it.
So we can considered that this is not an important warning in this case.

Patch-set: 5