Update patch set 1
Patch Set 1:

Noticed with the new factory image releases for 4.3 that a discrepancy exists between the certs used to sign the various system apps and those listed in the mac_permissions.xml file. Not sure on how you guys generate that final set of images, but we made a small adjustment to the post build scripts to properly replace the certs in the mac_permissions.xml file. Even if you don't use this set of tools, this change is warranted to maintain consistency for those that do.

Patch-set: 1
Reviewer: Robert Craig <1013157@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0