Update patch set 1
Patch Set 1: Code-Review+1

The symptom by the way of having the wrong certs in the mac_permissions.xml file is that all apps other than those running in fixed UIDs (e.g. system, radio, nfc) will be running in untrusted_app in the ps -Z output.  Whereas they should be in platform_app or release_app or media_app or shared_app if signed with one of the build keys.

Patch-set: 1
Label: Code-Review=+1