Update patch set 1
Patch Set 1: Code-Review-1

We intentionally made that be an error for sdk, for we want to make sure the SDK build installs everything in its product config.
Actually we want to enforce this check for all device builds as well, but unfortunately it takes much work to clean up the legacy product configs.
You have to build the sdk in the sdk product config; or you should clean the dangling module names from the product config.

Patch-set: 1
Reviewer: Ying Wang <1003981@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1