Update patch set 1
Patch Set 1:

I'm not entirely happy with the duplication here, but not sure how to avoid it and seems similar to the duplication we have among qcom-based board sepolicy files.  Especially as there are differences among some of the policy files.  I think we'll have to also add these for all of the generic_* boards here.
Alternatively, we can just not move this policy out of external/sepolicy; it doesn't technically cause any harm just a few extraneous domains/types.

Patch-set: 1