Update patch set 1
Patch Set 1:

> This patch needs the new 3.10 kernel with the 64bit ABI enabled.
 > Jun/David, were the latest binder patches from common merged into
 > the emulator kernel?

As Octavian's comments, a serials of goldfish binder patches need to be merged. It's still pending now. The userland binder should align to the goldfish kernel as the change in this one:
https://android-review.googlesource.com/#/c/81846/1/drivers/staging/android/uapi/binder.h

However, the goldfish kernel and system image in SDK are release together in a bundle. Therefore, this change will not impact the images were already released in the SDK.

Patch-set: 1
Reviewer: Jun Tian <1006838@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0