Update patch set 2
Patch Set 2:

As a bit of backstory, I'm running a fresh mavericks 10.9.2 install with an initial oracle jdk 1.7 package installation and no prior jdk installed. In this case, I do not have a symlink. However I tested using symlinks as well to verify this should work in different configurations. This patch should work in symlinked cases as well.

Initially, I was using conditional L flag to determine if /usr/libexec/java_home is a symlink, then readlink to resolve if so, but it was rather clunky. I've found we can grep `/usr/libexec/java_home`/release instead for a much cleaner solution.

As a minor note to a change I had a hard time justifying; the /usr/libexec/java_home evaluation is outside the darwin cases to save a few additional lines (because of the EXPERIMENTAL_USE_JAVA7 duplication).

Patch-set: 2