Update patch set 2
Patch Set 2:

> This going to force everything to rebuild, [...]

Nope, quite the opposite actually, it will force the rebuild of C/C++/ASM objects (as plugged in binary.mk) of the modules where the compilation flags change between builds (and only those), the others which do not change will not be rebuilt. 

That's what should be expected, because if the compilation flags (including defines) change between variants, the corresponding objects need to be rebuilt (otherwise the result of the build is actually undetermined and depends on what was built previously, which is not good).

 > so why not rm -rf $OUT/obj between targets?

Removing the objects as suggested (something similar to "make installclean") would force the rebuild even if not needed.

 > We've discussed this idea with the build
 > team before, and they have not been in favor of introducing this
 > many dependencies that would force rebuilds.

I understand and agree that we introduce more dependencies (each object depends on a compilation flags file, one per module), yet Make is precisely designed to manage dependencies between targets to ensure that the recipe is reapplied whenever needed and the result remains consistent. Which is expected from the build system.

Patch-set: 2
Reviewer: Olivier Fourdan <1042669@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0