Update patch set 1
Patch Set 1:

I successfully did a checkbuild with this today (plus a few LOCAL_GCCs added--dalvik, libdwfl, and bionic/test). Haven't tested on Mac which definitely needs to happen prior to submit, but I am definitely looking for feedback on whether or not this approach is reasonable.

Patch-set: 1
Label: Verified=0