Update patch set 5
Patch Set 5: Verified-1

This is an automated message from the Deckard Autoverifier.

Patchset merges but does not build in AOSP.
The following snippet of `make` output may point you in the right direction:

  build/core/target_test_internal.mk:18: *** frameworks/native/opengl/tests/hwc: Do not set LOCAL_MODULE_PATH when building test hwcStress.  Stop.



If your patchset has a dependency in a separate project, please ensure that
both changes are uploaded with the same Change-Id.

Patch-set: 5