Update patch set 2
Patch Set 2: Verified-1

This is an automated message from the Deckard Autoverifier.

Patchset merges but does not build in AOSP.
The following snippet of `make` output may point you in the right direction:

  make: *** No rule to make target `prebuilts/clang/linux-x86/host/3.5/bin/tblgen', needed by `out/host/linux-x86/gen/STATIC_LIBRARIES/libLLVMWrap_intermediates/llvm/IR/Intrinsics.gen'.  Stop.



If your patchset has a dependency in a separate project, please ensure that
both changes are uploaded with the same Change-Id.

Patch-set: 2