Update patch set 1
Patch Set 1:

> I think build-art-host is only used by build servers, hence not
 > knowing or caring about it. You should build ART from the ART
 > directory as presumably you don't want a minimal thing, but
 > something you can boot on a phone.

I'm doing this for a build server, hence this CL :-) The not knowing/caring part of your comments is scary though...

This CL is for running test-art-host 64bits on a build server. It's clearly not the way to go because Dalvik's still in our repo. I'll find an alternative, thanks for the comments, and thanks Ying for having tried the CL!

Patch-set: 1
Reviewer: Nicolas Geoffray <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0