Update patch set 1
Patch Set 1:

the -Bsymbolic dates from the initial import, so i doubt it was _done_ for security. i'd be interested to hear what the security guys think though. (bearing in mind that at the moment we have a non-RELRO public symbol for a global pointer that lets you intercept all the malloc functions :-) )

note that our choices are (at least):
* no default -Bsymbolic.
* no -Bsymbolic for libc.
* no -Bsymbolic (--dynamic-list) for the malloc functions in libc.
* public pointer to all the malloc functions. (status quo.)
* break asan.

Patch-set: 1
Label: Code-Review=0