Update patch set 1
Patch Set 1:

I didn't see USE_CLANG_PLATFORM_BUILD before. Looking at it now, my main issue with it is that it sets my_clang := true after the check for LOCAL_CLANG -- this makes it impossible to use clang as the default compiler, but force LOCAL_CLANG:=false in a couple of directories that are known troublemakers.

Any reason not to move the ifeq ($(USE_CLANG_PLATFORM_BUILD),true) block a couple of lines up so it comes before my_clang := $(strip $(LOCAL_CLANG))?

Patch-set: 1