Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I don't understand the motivation here. Both paths need to change honestly if you are going to do anything meaningful with a different set of prebuilts. It honestly feels a bit fragile to have part of it with a hard-coded "3.5", while you are seemingly trying to make it more generic.

Patch-set: 1
Label: Code-Review=-1