Update patch set 5
Patch Set 5:

I don't like it either, but this is the debate we had on the email thread before, right? There'd be something like 40 projects to update in master, plus an unknown number of unbundled projects. I can't really take on that much cleanup on my own.

Is there a way I could refactor the check into a common place?

Patch-set: 5
Label: Verified=0