Update patch set 2
Patch Set 2:

> I have been hesitant to add such dependency, because it usually
> means lots of unnecessary rebuild when makefile change. For
> example, if frameworks/base/Android.mk is modified, almost all Java
> code will be recompiled with this change.

This could be scaled back more if need be. We could remove it from the java.mk and friends (I'm not familiar with the Java side of our build system, so I don't know if it's necessary there). I think the important ones are binary.mk and (host_)?shared_library_internal.

Patch-set: 2