Update patch set 1
Patch Set 1:

I'll try this on darwin to see if the m option is present.

I'm not sure your example is a problem. Each static library is processed one at a time. The only problem is if a single library has the same named object file. I can test to be sure of this, though.

Patch-set: 1