Update patch set 9
Patch Set 9:

> > do you really need pattern matching for zipalign ? won't it be
 > > easier for the build system (or wrapper script) to expand
 > wildcards
 > > instead ?
 > 
 > I believe it will be less clearer in build system; but I do not
 > really know how to write makefiles.
 > 
 > > I'm just trying to avoid more code in zipalign, which
 > > suffers from a distinct lack of unit tests to begin with...
 > 
 > I think they are better of in zipalign. I can add some unit test to
 > as a separate change...

For an apk built from source, yes the build system knows what so files will be added to the apk;
But for a prebuilt apk, the build system will need to get the information from the prebuilt itself for example by running "unzip -l".
I prefer to doing it in zipalign to avoid this extra complication.

Patch-set: 9