Update patch set 1
Patch Set 1:

Last time I tried this (a few weeks ago, coincidentally), there were problems booting the system that I wasn't able to trace down. I'm not opposed to this, but I don't believe this will work as-is without fixing some other related code.

Even if we decide to accept this, this should be submitted as multiple different patches, one patch for each architecture, so that the patches can be rolled back individually as needed. Trying to do it for all architectures in one big shot isn't going to fly.

Patch-set: 1