Update patch set 1
Patch Set 1:

We're trying to enable the fast transition from COS to MOS and there is an issue with userdebug builds.

On a userdebug build the console is not started after the fast transition from COS to MOS while it is working just fine if we directly boot MOS.  The idea behind this patch is to use /system/default.prop to start console during the fast transition.  We thought it makes sense because of the obvious console service dependency on /system/bin/sh.

Any idea how we could fix this ?

Patch-set: 1