Update patch set 1
Patch Set 1:

Please note that this duplicates some of the -mfpu= work added by Bero.

Please note that this a request for comments patch and it might be heading in the totally wrong direction with 32bit ARMv8. We definitely do not want 32bit NDK apps that are built only for ARMv8 AArch32, but we should, if possible, make use of the ARMv8 features in the platform side.

Please take a look at my comments in core/combo/arch/arm/armv8-a.mk. There are still issues related to building the codebase with -march=armv8-a or -mcpu=[ARMv8 CPUs]

Patch-set: 1