Update patch set 2
Patch Set 2:

The build servers also do FORCE_BUILD_LLVM_COMPONENTS=true to work around this.

One thing that we had discussed in the past was rebuilding libclang.so and the like with instrumentation so that this wasn't necessary (and then SANITIZE_HOST=address would be the default). Given that we might want to do other sanitizers on the host though, I don't know if that makes sense any more. Discussion was on http://b/18652305

I suppose an alternative would be to make a lib/asan/libclang.so for the host.

Patch-set: 2
Reviewer: Dan Albert <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0