Update patch set 1
Patch Set 1:

(1 comment)

Yeah, this primarily was created because of the androidbp conversion, but it seems to be a pretty common use case -- 29% of the libraries in soong use cc_library (28/98). Not all of those expose both static and shared libraries in the make build system.

In the make build system I see these potential users (on AOSP):

  bootable/recovery/minui
  external/e2fsprogs/lib/blkid
  external/e2fsprogs/lib/e2p
  external/e2fsprogs/lib/et
  external/e2fsprogs/lib/ext2fs
  external/e2fsprogs/lib/quota
  external/e2fsprogs/lib/uuid
  external/jpeg
  external/libnl
  external/libsepol
  external/libxml2
  external/mdnsresponder
  external/selinux/libsepol
  hardware/libhardware_legacy
  hardware/ril/librilutils
  packages/inputmethods/LatinIME/native/jni
  system/core/base
  system/core/libcutils
  system/core/liblog
  system/core/libsparse
  system/core/libsuspend
  system/core/libusbhost
  system/core/libutils
  system/core/libziparchive
  system/core/logwrapper
  system/extras/ext4_utils
  system/extras/f2fs_utils

Patch-set: 1
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0