Update patch set 1
Patch Set 1:

> As with anything, no free lunch. More memory, less cycles, yada
 > yada.
 > 
 > My only concern is size. It might take people a while to get off of
 > ramdisks, and I dont have the data handy, but what happens to
 > Samsung's file_contexts, how big do they get? However, if we get
 > rid of the ramdisk as Nick mentioned, that frees up a lot of space
 > for this. Although just becuase their is more space doesnt mean we
 > should consume it. Also, if more userspace processes use
 > libselinux, they will bloat as well.
 > 
 > Has their been any major issues with performance on these execution
 > paths?

In Linux distributions, yes, which is why file_contexts.bin was introduced upstream.  Not sure about Android, although Richard measured the speed-up from file_contexts.bin as ranging from 2X for 500 entries (ala Android) to 5X for 6000 entries (ala Linux).  Not sure though how that stands relative to overall boot time on Android.

It shouldn't increase memory usage by the processes; it is simply a precompiled form of file_contexts, so they would still be using the same memory.  It just increases storage cost.

To fully remove file_contexts (which is not done in Linux distributions, although it is only loaded if file_contexts.bin does not exist, is older, or cannot be loaded successfully, e.g. pcre version mismatch), we would need to change libselinux (as it currently tries opening and stat'ing file_contexts first and then compares the timestamps, and fails if it does not exist).  We would also need to update the build makefiles and releasetools for using file_contexts.bin, and the CTS (including a way to compare two file_contexts.bin files).  So it isn't completely trivial.

Patch-set: 1