Update patch set 1
Patch Set 1:

But that's only for really special modules. We don't have any today, and if we add TSan, there will be 1 or 2 (corresponding tests). Actually, even then we won't need this if we just don't build anything that depends on those targets (and they won't be in the image so that's easy to achieve).

So, yes, make SANITIZE_TARGET=address override thread and merge with anything else.

Patch-set: 1
Reviewer: Evgenii Stepanov <1005967@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0