Update patch set 1
Patch Set 1:

I start to wonder if this is not becoming too difficult to understand. For 32bit ARMv8 (Denver, a53, a57, ..) we either build with -mcpu=cortex-a15 or -mcpu=cortex-a7 (as of this patch).

This is mostly because if we use -mcpu=cortex-a5X for the 32bit side that will also enable ARCH_ARM > 7, which previously was not acceptable.

Would it be better if we accept this for now and we start fixing things for the platform so that a 32bit cortex-a53 is built with -mcpu=cortex-a53 ? 

My only worry is doing this in the NDK. I really don't  think we should support a new 32bit ABI there. But maybe we can fix that with some:

#if NDK_BUILD
#if ARCH_ARM > 7
#error "Not supported"
#else /*Android platform */
...

Patch-set: 1