Update patch set 1
Patch Set 1:

This doesn't mean that the module that requires safe stack is actually installed -- it doesn't even have to be built for this to be set.

Is that the expected behavior, or do you only want it set only if a module that requires it is installed (that's much harder).

Patch-set: 1
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0