Update patch set 1
Patch Set 1:

> This patchset was reverted in change: I925dae6a2d07792c61fcd3b6fd8718391ab993f8

This change breaks the FrameworksCoreTests build, because it has a resource file name org/apache/harmony/tests/java/lang/test#.properties, for which jar mistakes # as the the line comment start mark.
We need to quote the arguments. A new change was uploaded: https://android-review.googlesource.com/175954

Patch-set: 1