Update patch set 3
Patch Set 3:

> arm/arm64 uses 'generic' instead of 'default' (at least in bionic).
 > x86 falls back to the 'generic' name in the makefiles too if the
 > variant isn't understood

These are ART/dex2oat variants, which are totally orthogonal. All our architectures support "default," but only arm64 does (where it's the same as "default"; don't know why that was introduced).

I also somewhat doubt that the build system understands all the variants ART knows (!= takes advantage of). Take a look into the ARM instruction-set-features, for example.

It was originally the intent of irogers to try to be as similar to C compilers with this, but the build system support never got as complicated. I can go and clean up our side in a follow-up, if you or Ying want me to...

Patch-set: 3